Refactor: reorganize source files into subdirectories
Moved core-related files into a new 'Core' directory, DebugWindow files into 'DebugWindow', and SelfieSegmenter into 'SelfieSegmentator'. Updated CMakeLists.txt and include paths accordingly to reflect the new structure. This improves project organization and maintainability.
