+++ src/Core/MainEffect.hpp
