+++ src/Core/MainPluginContext.cpp
