+++ src/Core/MainPluginContext.h
