+++ src/Core/RenderingContext.cpp
